feat(clover): add correct doclinks for hetzner
